Move services to /services
