+++ src/app/app.module.ts
@@ -4,7 +4,7 @@ import { NgModule } from '@angular/core';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
-import { InterceptorService } from './auth/interceptor.service';
+import { InterceptorService } from './services/interceptor.service';
 import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
 
 @NgModule({
