+++ src/app/auth/auth.guard.spec.ts
@@ -2,7 +2,7 @@ import { TestBed } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
 
 import { AuthGuard } from './auth.guard';
-import { AuthService, AuthServiceMock } from './auth.service';
+import { AuthService, AuthServiceMock } from '../services/auth.service';
 
 describe('AuthGuard', () => {
   let guard: AuthGuard;
