+++ src/app/auth/register/register.component.spec.ts
@@ -1,6 +1,6 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
-import { AuthService, AuthServiceMock } from '../auth.service';
+import { AuthService, AuthServiceMock } from '../../services/auth.service';
 
 import { RegisterComponent } from './register.component';
