+++ src/app/authors/authors.component.spec.ts
@@ -1,9 +1,9 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
-import { AuthService, AuthServiceMock } from '../auth/auth.service';
+import { AuthService, AuthServiceMock } from '../services/auth.service';
 
 import { AuthorsComponent } from './authors.component';
-import { AuthorsService, AuthorsServiceMock } from './authors.service';
+import { AuthorsService, AuthorsServiceMock } from '../services/authors.service';
 
 describe('AuthorsComponent', () => {
   let component: AuthorsComponent;
