+++ src/app/authors/authors.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { faSync, faPlus } from '@fortawesome/free-solid-svg-icons';
-import { AuthService } from '../auth/auth.service';
-import { AuthorsService } from './authors.service';
+import { AuthService } from '../services/auth.service';
+import { AuthorsService } from '../services/authors.service';
 
 @Component({
   selector: 'app-authors',
