+++ src/app/books/book-details/book-details.component.spec.ts
@@ -1,8 +1,8 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
-import { AuthService, AuthServiceMock } from 'src/app/auth/auth.service';
-import { AuthorsService, AuthorsServiceMock } from 'src/app/authors/authors.service';
-import { BooksService, BooksServiceMock } from '../books.service';
+import { AuthService, AuthServiceMock } from '../../services/auth.service';
+import { AuthorsService, AuthorsServiceMock } from '../../services/authors.service';
+import { BooksService, BooksServiceMock } from '../../services/books.service';
 
 import { BookDetailsComponent } from './book-details.component';
