+++ src/app/users/users.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { faSync, faPlus } from '@fortawesome/free-solid-svg-icons';
-import { AuthService } from '../auth/auth.service';
-import { UsersService } from './users.service';
+import { AuthService } from '../services/auth.service';
+import { UsersService } from '../services/users.service';
 
 @Component({
   selector: 'app-users',
